add hot reload feature to vite and rollup

+++ packages/cppjs-plugin-rollup/index.js
@@ -1,9 +1,11 @@
 import CppjsCompiler from 'cpp.js';
 
 import fs from 'fs';
+import p from "path";
 
 const rollupCppjsPlugin = (options, _compiler) => {
     const compiler = _compiler || new CppjsCompiler(options);
+    const headerRegex = new RegExp(`.(${compiler.config.ext.header.join('|')})$`);
 
     return {
         name: 'rollup-plugin-cppjs',
@@ -14,13 +16,33 @@ const rollupCppjsPlugin = (options, _compiler) => {
             return null;
         },
         async transform(code, path) {
-            if (!/.h$/.test(path)) {
-                return
+            if (!headerRegex.test(path)) {
+                return;
             }
 
             compiler.findOrCreateInterfaceFile(path);
 
             return "export default function() { return new Promise((resolve, reject) => import('/cpp.js').then(n => n.default(resolve))); }";
+        },
+        buildStart() {
+            const watch = (dirs) => {
+                dirs.forEach(dir => {
+                    const filesToWatch = fs.readdirSync(dir);
+
+                    for (let file of filesToWatch) {
+                        const fullPath = p.join(dir, file);
+                        const stats = fs.statSync(fullPath);
+
+                        if (stats.isFile()) {
+                            this.addWatchFile(fullPath);
+                        } else if (stats.isDirectory()) {
+                            watch([fullPath]);
+                        }
+                    }
+                });
+            };
+
+            watch(compiler.config.paths.native);
         },
         generateBundle() {
             compiler.createBridge();
@@ -35,7 +57,10 @@ const rollupCppjsPlugin = (options, _compiler) => {
                 source: fs.readFileSync(`${compiler.config.paths.temp}/${compiler.config.general.name}.wasm`),
                 fileName: "cpp.wasm"
             });
-            fs.rmSync(compiler.config.paths.temp, { recursive: true, force: true });
+            const isWatching = process.argv.includes('-w') || process.argv.includes('--watch');
+            if (!isWatching) {
+                fs.rmSync(compiler.config.paths.temp, { recursive: true, force: true });
+            }
         },
     }
 };
